Feed Feature - Added FeedItem model and FeedLoader protocol
1.To get Feed, we need to create a model which contains all required data like its id, description, location, image URL. That is why a model is created
2.
a.To get Feed, we need a method which does this work. That is why  load() method is created and as fetching is time consuming we implement it in other thread and as its implemented in other use escaping closure to return ur value.

b.In closure, we expect to get return value as FeedItem but as fetching feeds might fail instead of output as FeedItem keep it as enum with two possible values ie., either FeedItem or Error