1.Added documentation files for Architechture, Network Layers
2.Refactored code in HTTPClient because what we really need is getting data from url and not setting the requested url

3.Also, setting requested url is not at all our requirement but still we need it for test cases to check if url exists or not when httpclient calls get data from url.
So instead of keeping it in production code ie., http client shift it to test code. Create separate spy, so it can be used to test if requestedUrl is set or not

a.To create a spy, we need a subclass. As our httpclient is private init, outside of it we cannot create it and thorugh shared property we can access http client. So to create subclass for sure private init should be removed from HTTPClient
b.In test case, we need to replace HTTPClient.shared with spy because in our RemoteFeedLoader we are using HTTPClient.shared. Try to understand that shared is just a property which can store any instance. by default it has HTTPClient() but we can replace it with spy as we need it to satisfy test case
Note: ofcourse, this is very bad code but first we need to satisfy our test case then we need refactor