1.As using concrete HTTPClient is anti pattern, removed its concrete object and passed it through dependency injection in load function.
a.Try to understand that even if it is passed as dependency injection, our purpose is served and at the same time we are eliminating tight coupling.
We also made it a normal class and removed it from being global shared instance.

Note: any class which uses "static var shared = instance", shares global mutatble state as shared property can be modified from anywhere in the app. So we might end up having differnet instance for "shared" property in different threads

b.As its not concrete object anymore, setting spy to httpclient.shared is not required anymore and still test cases passes perfectly