1.As HTTPClient class just acts an abstract class ie., it does not have any implementation inside may be it looked as if we can replace it with protocol which makes code more simpler
2.Refactored code to include all configuration at single place in makeSut, used appropriate names for tuple return type so we can avoid referencing with 0, 1
set url's default value because not all test cases needs it mandatorily

3.Along with client, passed url also in DI through constructor DI for Remote Feed Loader.

a.Before we hardcoding URL within load function. If we hardcode it then this function will be useful only for this one Server. What if we need to facilitate same code for deve, staging, prid differently then we need to add if else conditions within this load function to declare those urls.

b.It violates Solid principles. Instead if we can pass url to load function then this function can be used for any environment. As we have only one remote feed loader instance, we instead pass it to constructor injection instead of method injection

c.But we have two options here - one is to pass url to HTTPClient's constructor injection and other is to pass to RemoteFeedLoader init